+++ app.js
@@ -34,7 +34,7 @@ app.use((error, req, res, next) => {
   //   return res.status(error.status || 500).json({ message: error.message || "Internal server error" })
   // }
   // console.error(error)
-  return res.status(400).json("Помилка валідації")
+  return res.status(400).json({message: error})
 });
 
 module.exports = { app, };
